Initial commit, added WkbConverter and EntityFrameworkConvert. Pending tests
