+++ src/GeoJSON.Net.Contrib.Wkb/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="GeoJSON.Net" version="1.0.62" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net452" />
+</packages>
