+++ test/GeoJSON.Net.Contrib.MsSqlSpatial.Test/Properties/AssemblyInfo.cs
