+++ test/GeoJSON.Net.Contrib.MsSqlSpatial.Test/WktSamples.cs
