+++ test/GeoJSON.Net.Contrib.MsSqlSpatial.Test/WktTests.cs
