+++ test/GeoJSON.Net.Contrib.MsSqlSpatial.Test/packages.config
