+++ test/GeoJSON.Net.Contrib.Wkb.Test/WkbConverterTests.cs
@@ -1,0 +1,12 @@
+﻿using Xunit;
+
+namespace GeoJSON.Net.Contrib.Wkb.Test
+{
+    public class WkbConverterTests
+    {
+        [Fact]
+        public void TestMethod1()
+        {
+        }
+    }
+}
